requirements.txt
PyMySQL
six>=1.7.0 # MIT
SQLAlchemy<1.1.0,>=0.9.9
kafka-python
simport>=0.0.dev0
stevedore>=1.5.0 # Apache-2.0
oslo.config>=1.2.1
oslo.log
oslo.service
tooz
